+++ src/com/maddyhome/idea/vim/group/CopyGroup.java
@@ -1,6 +1,6 @@
 /*
  * IdeaVim - Vim emulator for IDEs based on the IntelliJ platform
- * Copyright (C) 2003-2014 The IdeaVim authors
+ * Copyright (C) 2003-2016 The IdeaVim authors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -439,9 +439,7 @@ public class CopyGroup {
         break;
     }
 
-    final MarkGroup markGroup = VimPlugin.getMark();
-    markGroup.setMark(editor, '[', offset);
-    markGroup.setMark(editor, ']', endOffset);
+    VimPlugin.getMark().setChangeMarks(editor, new TextRange(offset, endOffset));
   }
 
   private static final Logger logger = Logger.getInstance(CopyGroup.class.getName());