+++ src/com/maddyhome/idea/vim/option/Options.java
@@ -1,6 +1,6 @@
 /*
  * IdeaVim - Vim emulator for IDEs based on the IntelliJ platform
- * Copyright (C) 2003-2014 The IdeaVim authors
+ * Copyright (C) 2003-2016 The IdeaVim authors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -19,8 +19,10 @@ package com.maddyhome.idea.vim.option;
 
 import com.intellij.openapi.diagnostic.Logger;
 import com.intellij.openapi.editor.Editor;
+import com.intellij.openapi.extensions.Extensions;
 import com.maddyhome.idea.vim.VimPlugin;
 import com.maddyhome.idea.vim.ex.ExOutputModel;
+import com.maddyhome.idea.vim.extension.VimExtension;
 import com.maddyhome.idea.vim.helper.EditorHelper;
 import com.maddyhome.idea.vim.helper.MessageHelper;
 import com.maddyhome.idea.vim.helper.Msg;
@@ -36,6 +38,7 @@ public class Options {
   public static final String RELATIVE_NUMBER = "relativenumber";
   public static final String NUMBER = "number";
   public static final String CLIPBOARD = "clipboard";
+  public static final String INCREMENTAL_SEARCH = "incsearch";
 
   /**
    * Gets the singleton instance of the options
@@ -463,6 +466,33 @@ public class Options {
     addOption(new ToggleOption(NUMBER, "nu", false));
     addOption(new ToggleOption(RELATIVE_NUMBER, "rnu", false));
     addOption(new ListOption(CLIPBOARD, "cb", new String[]{"autoselect,exclude:cons\\|linux"}, null));
+    addOption(new ToggleOption(INCREMENTAL_SEARCH, "is", false));
+
+    registerExtensionOptions();
+  }
+
+  private void registerExtensionOptions() {
+    for (VimExtension extension : Extensions.getExtensions(VimExtension.EP_NAME)) {
+      final String name = extension.getName();
+      final ToggleOption option = new ToggleOption(name, name, false);
+      option.addOptionChangeListener(new OptionChangeListener() {
+        @Override
+        public void valueChange(OptionChangeEvent event) {
+          for (VimExtension extension : Extensions.getExtensions(VimExtension.EP_NAME)) {
+            if (name.equals(extension.getName())) {
+              if (Options.getInstance().isSet(name)) {
+                extension.init();
+                logger.info("IdeaVim extension '" + name + "' initialized");
+              }
+              else {
+                extension.dispose();
+              }
+            }
+          }
+        }
+      });
+      addOption(option);
+    }
   }
 
   private void addOption(@NotNull Option option) {